Update k8s-topgun for Helm v3
Behaviour around namespaces changed in Helm 3. Every helm command is now
scoped to a namespace just like kubectl. We now need to specify the
namespace in each helm command.

Namespaces are also not made automatically by helm if they don't exist.
This results in "Namespace does not exist" errors from helm when trying
to install a chart. This commit adds a function to create the namespace
right before installing a chart using kubectl. There's a proposal to
bring this behaviour back into helm:
https://github.com/helm/helm/issues/6794

Tiller is no longer required either, the `init` command was removed
because of this.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>