atc: behaviour: implement ListAllJobsWatcher
concourse/rfcs#61

This commit enables monitoring the relevant database tables for watching
changes to the ListAllJobs endpoint. The `ListAllJobsWatcher` is roughly
as described in the linked RFC, in that it uses Postgres TRIGGERs +
NOTIFY/LISTEN to capture/observe changes to the relevant tables. Using
the primary keys from these tables, it then performs the ListAllJobs
query scoped to only the affected jobs. It will then send the updated
job(s) to each subscriber that has access to them

Currently, we aren't taking into account public pipelines for access
control - this'll be addressed in a later commit.

One thing to note is that the list of tables/columns being monitored for
changes is kind of arbitrary, and it's possible it won't capture every
change to jobs. I initially started by monitoring changes to any field
from any table that is referenced in the ListAllJobs queries. This would
work, but would also result in duplicate events. It seems like most
changes to jobs seem to flow through the jobs table at some point (often
via the `config`).

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>